requirements.txt
openai
python-dotenv
requests
yfinance

#pip install -r requirements.txt